requirements.txt
Flask>=3.0,<4.0
gunicorn
itsdangerous>=2.2
PyJWT>=2.8
psycopg2-binary
SQLAlchemy>=2.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
alembic
redis>=5.0
qrcode[pil]
python-dotenv
requests
user-agents
